% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat-disp.R
\name{stat_disp}
\alias{stat_disp}
\title{Title}
\usage{
stat_disp(mapping = NULL, data = NULL, geom = "polygon",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{geom}{}

\item{position}{}

\item{na.rm}{}

\item{show.legend}{}

\item{inherit.aes}{}

\item{...}{}
}
\value{
1
}
\description{
Title
}
\examples{
1
}

