% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat-disp.R
\docType{data}
\name{StatDisp}
\alias{StatDisp}
\title{Title}
\format{\preformatted{Classes 'StatDisp', 'Stat', 'ggproto' <ggproto object: Class StatDisp, Stat>
    compute_group: function
    compute_layer: function
    compute_panel: function
    default_aes: uneval
    extra_params: na.rm
    non_missing_aes: 
    parameters: function
    required_aes: x y fill
    retransform: TRUE
    setup_data: function
    setup_params: function
    super:  <ggproto object: Class Stat> 
}}
\usage{
StatDisp
}
\arguments{
\item{data}{}

\item{scales}{}

\item{params}{}

\item{fx}{}

\item{fy}{}

\item{nx}{}

\item{ny}{}

\item{method}{}

\item{midpoint}{}
}
\value{
1
}
\description{
Title
}
\examples{
1
}
\keyword{internal}

